fix to clang error

+++ exo/inference/tinygrad/models/llama.py
@@ -253,6 +253,6 @@ def fix_bf16(weights: Dict[Any, Tensor]):
     if Device.DEFAULT == "CLANG":
         # TODO: Without casting to float16, 70B llama OOM on tinybox.
         return {
-          k: (v.llvm_bf16_cast(dtypes.float32).to(v.device) if v.dtype == dtypes.bfloat16 else v) for k, v in weights.items()
+          k: v.llvm_bf16_cast(dtypes.half).to(v.device) if v.dtype == dtypes.bfloat16 else v for k, v in weights.items()
         }
     return {k:v.cast(dtypes.float16) if v.dtype == dtypes.bfloat16 else v for k,v in weights.items()}
